feat(tokenize-cli): add simple tokenize CLI to test and experiment
The CLI simply fetches Messages as Observables and passes them through the
one-step chain of elaboration.
Future CLIs will use an abstraction of this one in order to eventually migrate to a
"playground-cli" module that will be able of composing tasks in a single chain.
This will prove useful when all tasks are going to be composed in one single
giant flow.